feat: changed id type and some other minor improvements

+++ domain/src/main/java/pt/uminho/dss/restaurante/domain/entity/Alergenico.java
@@ -1,7 +1,7 @@
 package pt.uminho.dss.restaurante.domain.entity;
 
 public class Alergenico {
-    public long id;
+    public Integer id;
     public String nome;
 
     // Construtores
@@ -14,8 +14,8 @@ public class Alergenico {
 
     // Getters e Setters
 
-    public long getId() { return id; }
-    public void setId(long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
