+++ domain/src/main/java/pt/uminho/dss/restaurante/domain/entity/LinhaMenu.java
@@ -3,7 +3,7 @@ package pt.uminho.dss.restaurante.domain.entity;
 import java.io.Serializable;
 
 public class LinhaMenu implements Serializable {
-    private Long id;
+    private Integer id;
     private Produto produto;
     private Integer quantidade;
 
@@ -18,8 +18,8 @@ public class LinhaMenu implements Serializable {
 
     // Getters e Setters
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public Produto getProduto() { return produto; }
     public void setProduto(Produto produto) { this.produto = produto; }
