+++ domain/src/main/java/pt/uminho/dss/restaurante/domain/entity/Stock.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 import java.util.List;
 
 public class Stock implements Serializable {
-    private Long id; // não sei se deve exister
+    private Integer id; // não sei se deve exister
     private List<LinhaStock> ingredientes;
 
     // Construtores
@@ -22,8 +22,8 @@ public class Stock implements Serializable {
 
     // Getters e Setters
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public List<LinhaStock> getIngredientes() { return ingredientes; }
     public void setIngredientes(List<LinhaStock> ingredientes) { this.ingredientes = ingredientes; }
